updated views and forms
