+++ apps/user/forms/__init__.py
@@ -1,0 +1,3 @@
+from .auth_forms import UserRegisterForm  # add others from auth_forms if any
+from .account_forms import EmailChangeForm, CustomPasswordChangeForm
+from .account_forms import DeleteAccountForm
