+++ apps/user/forms/auth_forms.py
@@ -1,0 +1,66 @@
+from django import forms
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.models import User
+from django.core.exceptions import ValidationError
+from django.contrib.auth.password_validation import (
+    password_validators_help_text_html,
+)
+
+
+# ----------------- Registration Form -----------------
+class UserRegisterForm(UserCreationForm):
+    """
+    Custom user registration form extending Django's UserCreationForm.
+    Adds email field with uniqueness check and Bootstrap styling.
+    """
+    username = forms.CharField(
+        label="Username",
+        max_length=50,
+        required=True,
+        widget=forms.TextInput(attrs={
+            'class': 'form-control',
+            'placeholder': 'Choose a username'
+        })
+    )
+
+    email = forms.EmailField(
+        label="Email",
+        max_length=100,
+        required=True,
+        widget=forms.EmailInput(attrs={
+            'class': 'form-control',
+            'placeholder': 'Enter your email'
+        })
+    )
+
+    password1 = forms.CharField(
+        label="Password",
+        widget=forms.PasswordInput(attrs={
+            'class': 'form-control',
+            'placeholder': 'Create a password'
+        }),
+        help_text=password_validators_help_text_html()
+    )
+
+    password2 = forms.CharField(
+        label="Confirm Password",
+        widget=forms.PasswordInput(attrs={
+            'class': 'form-control',
+            'placeholder': 'Confirm your password'
+        })
+    )
+
+    class Meta:
+        model = User
+        fields = ['username', 'email', 'password1', 'password2']
+
+    def clean_email(self):
+        """
+        Ensure the email is unique across all users.
+        """
+        email = self.cleaned_data.get("email")
+        if User.objects.filter(email=email).exists():
+            raise ValidationError(
+                "A user with this email already exists."
+            )
+        return email
